requirement.txt
#torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
accelerate==0.16.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bottleneck==1.4.2
cattrs==24.1.2
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.0
coremltools==8.1
cycler==0.12.1
easydict==1.13
einops==0.8.0
ema-pytorch==0.7.7
eval-type-backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.12.23
flet==0.25.2
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
graphviz==0.20.3
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iopath==0.1.10
jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.4
kornia-rs==0.1.7
lazy-loader==0.4
lmdb==1.5.1
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.4
multidict==6.1.0
networkx==3.2.1
numexpr==2.10.1
numpy==2.0.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openjpeg==2.5.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.1.1
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyaml==24.12.1
pydantic==2.10.4
pydantic-core==2.27.2
pyparsing==3.2.1
pysocks==1.7.1
python-dateutil==2.9.0post0
python-tzdata==2023.3
pytorch-lightning==1.9.3
pytz==2024.2
pyyaml==6.0.2
qudida==0.0.4
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20250102
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
theano==1.0.5
thop==0.1.1-2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.6.12
tomli==2.2.1
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.6.1
torchvision==0.16.0
torchviz==0.0.3
tqdm==4.67.1
transparent-background==1.3.3
typing_extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
wget==3.2
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0